Create InputDateTime.js
Alternative to the DatePicker and TimePicker.  This version uses an HTML5 <input type="date"> (or "time").  The latest versions of iOS and Android display a native picker when this is selected.  Since the date and time types are not styled in WinJS, this custom control uses an <input type="text"> as a visual placeholder for the hidden date or time input.